[CHORE] #Voice - 데이터 전송 테스트 함수 분리, 재명명
