+++ src/pages/voice-recognition/TextPage.jsx
@@ -6,7 +6,7 @@ import Pen from '@/assets/images/Pen.svg'
 import { useState } from 'react';
 import Loading from '../../components/Loading/Loading';
 import { useNavigate } from 'react-router-dom';
-import { getRouletteDataTest } from '../../api/ApiClient';
+import { getKeywordsTest } from '@/api/getKeywords.jsx';
 
 const TextPage = () => {
   const navigate = useNavigate();
@@ -16,8 +16,8 @@ const TextPage = () => {
   const handleSubmit = async () => {
     setIsLoading(true);
     setTimeout(() => {
-      getRouletteDataTest(content, navigate);
-    //getRouletteData(transcript, navigate);
+      getKeywordsTest(content, navigate);
+    //getKeywords(content, navigate);
   }, 2000);
 }
 return (
